Separate timing functionality from BaseAgent
