+++ src/main/java/org/maas/Start.java
@@ -1,13 +1,12 @@
-package org.yourteamname;
+package org.maas;
 
 import java.util.List;
 import java.util.Vector;
-import org.yourteamname.agents.DummyAgent;
 
 public class Start {
     public static void main(String[] args) {
     	List<String> agents = new Vector<>();
-    	agents.add("test:org.yourteamname.agents.DummyAgent");
+    	agents.add("test:org.maas.agents.DummyAgent");
 
     	List<String> cmd = new Vector<>();
     	cmd.add("-agents");
