+++ src/main/java/org/maas/agents/DummyAgent.java
@@ -1,4 +1,4 @@
-package org.yourteamname.agents;
+package org.maas.agents;
 
 import jade.content.lang.Codec;
 import jade.content.lang.sl.SLCodec;
